Add 1raster converter, build scripts, and fixes
Add tools and platform build support and apply numerous safety and feature fixes. New files: .gitignore, CHANGELOG.md, docs/gtultra-to-1raster.md, macOS/MSYS2/Windows build scripts, tools/gtultra2raster converter and assets, and gtasm wrapper. Update README with build and converter instructions. Multiple source fixes and improvements: remove debug spew and disable yydebug in the assembler, use %p for pointer logging, initialize local variables and byte buffers, validate and clamp I/O reads, fix endianness casts, correct sentinel comparisons, add bounded string appends to avoid overflows, and improve palette-name allocation. Add relocator ASM output support (FORMAT_ASM), embed author-info patching, safer SID address validation/syncing, and adjust relocator I/O to write text mode for ASM. Miscellaneous makefile and build rule tweaks to support macOS/Windows/MSYS2 builds and avoid stale object-file cross-contamination.

+++ src/resid-fp/sidfp.cpp
@@ -19,6 +19,7 @@
 
 #include "sidfp.h"
 #include <math.h>
+#include <stdint.h>
 
 #define RINGSIZE 2048
 
@@ -708,10 +709,9 @@ static float convolve(const float *a, const float *b, int n)
 #ifdef __SSE__
 	__m128 out4 = { 0, 0, 0, 0 };
 
-	/* examine if we can use aligned loads on both pointers -- some x86-32/64
-	 * hackery here ... should use uintptr_t, but that needs --std=C99... */
+	/* examine if we can use aligned loads on both pointers */
 	int diff = static_cast<int>(a - b) & 0xf;
-	int a_align = static_cast<int>(reinterpret_cast<long>(a)) & 0xf;
+	int a_align = static_cast<int>(reinterpret_cast<uintptr_t>(a)) & 0xf;
 
 	/* advance if necessary. We can't let n fall < 0, so no while (n --). */
 	while (n > 0 && a_align != 0 && a_align != 16) {
